Move command line to new main module
